refactor(entries): data-table の責務をフック・サブコンポーネントに分離
- useEntryTable: テーブル状態・データ正規化・フィルタ・セル編集
- useBulkDeleteEntries: 一括削除フロー
- BulkDeleteEntriesDialog: 削除確認UI
- normalize-entries: エントリ正規化ロジック
- data-table.tsx を591行から164行に縮小（配線中心）

Refs: #134
